Exercise_9 new variable names: corrected.

+++ week_02/day_1/exercise_9.cpp
@@ -13,7 +13,7 @@
 using namespace std;
 
 int main() {
-	int i = 1357988018575474;
+	long long i = 1357988018575474;
 	// tell if it has 11 as a divisor
 
 	if (i % 11 == 0){
